:recycle: use context instead of redux
