+++ README.en.md
@@ -19,8 +19,8 @@ Cloud air conditioner. Portable air conditioner. Invite a cool breeze into your 
 
 > History: [云空调，便携小空调｜云游君的小站](https://www.yunyoujun.cn/posts/air-conditioner/)
 
-- Machine Only：[ac.yunyoujun.cn](https://ac.yunyoujun.cn)
-- Test Machine：[ac.yyj.moe](https://ac.yyj.moe)
+- Machine Only[master]：[ac.yunyoujun.cn](https://ac.yunyoujun.cn)
+- Test Machine[dev]：[ac.yyj.moe](https://ac.yyj.moe)
 - Sample Room：<https://www.yunyoujun.cn/air-conditioner-room/>
 
 ## Features
