+++ package.json
@@ -37,32 +37,29 @@
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
     "@mui/material": "^5.6.1",
-    "@reduxjs/toolkit": "^1.8.1",
     "react": "17",
     "react-dom": "17",
     "react-ga": "^3.3.0",
     "react-gtm-module": "^2.0.11",
-    "react-redux": "^7.2.8",
     "react-router-dom": "^6.3.0",
     "sass": "^1.50.0",
     "web-vitals": "^2.1.4"
   },
   "devDependencies": {
     "@antfu/eslint-config": "^0.20.2",
+    "@iconify-json/ic": "^1.1.2",
     "@iconify-json/mdi": "^1.1.9",
-    "@mui/icons-material": "^5.6.1",
     "@svgr/core": "^6.2.1",
     "@types/node": "^17.0.23",
     "@types/react": "17",
     "@types/react-dom": "17",
     "@types/react-gtm-module": "^2.0.1",
-    "@types/react-redux": "^7.1.23",
     "@types/react-router-dom": "^5.3.3",
     "@vitejs/plugin-react": "^1.3.0",
     "eslint": "^8.13.0",
     "typescript": "^4.6.3",
     "unocss": "^0.31.0",
-    "usehooks-ts": "^2.5.0",
+    "usehooks-ts": "^2.5.1",
     "vite": "^2.9.1",
     "vite-plugin-pages": "^0.22.0",
     "vite-plugin-pwa": "^0.11.13"
