+++ src/App.scss
@@ -1,8 +1,3 @@
-@font-face {
-  font-family: "Digital-7 Mono";
-  src: url("/assets/fonts/digital-7-mono.ttf") format("truetype");
-}
-
 .hot-color {
   filter: saturate(110%);
 }
