--- src/app/store.ts
@@ -1,20 +1,0 @@
-import type { Action, ThunkAction } from '@reduxjs/toolkit'
-import { configureStore } from '@reduxjs/toolkit'
-import acReducer from '../features/ac/acSlice'
-import toastReducer from '../features/toast/toastSlice'
-
-export const store = configureStore({
-  reducer: {
-    ac: acReducer,
-    toast: toastReducer,
-  },
-})
-
-export type AppDispatch = typeof store.dispatch
-export type RootState = ReturnType<typeof store.getState>
-export type AppThunk<ReturnType = void> = ThunkAction<
-ReturnType,
-RootState,
-unknown,
-Action<string>
->
