+++ src/pages/Rc.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import { RemoteControl } from '~/components/RemoteControl'
-import { useDetectStorage } from '~/features/ac'
+import RemoteControl from '~/components/RemoteControl'
+import { useDetectStorage } from '~/hooks'
 
 const Rc: React.FC = () => {
   useDetectStorage()
