+++ src/styles/helper.scss
@@ -1,0 +1,12 @@
+@font-face {
+  font-family: "Digital-7 Mono";
+  src: url("/assets/fonts/digital-7-mono.ttf") format("truetype");
+}
+
+.font-digit {
+  font-family: "Digital-7 Mono";
+}
+
+.ac-text {
+  color: var(--ac-text-color);
+}
